Remove preferences.
We decided that we don't really need a prompt for querying the agent,
since installation of extension already indicates desire to use
monkeysphere.  Given that this was really the only preference being
used at the moment, I remove all the preference infrastructure here,
in the name of simplicity.  If we need preferences down the line we
can re-add.