Implement cache.
The cache is set on response from the agent.  The cache is queried
onLocationChange to see if the status bar needs to be updated.  dkg's
MITM ocncerns might still need to be addressed, but this gives a
working framework.  issues:

* status is not updated after agent query for NOTVALID sites.  don't
  know where to put in an updateStatus hook, because just reloading
  will put the site into a reload loop

* switching to a blank tab will leave any monkeysphere status icons in
  the status bar, because i can't figure out what listener is
  triggered on switching to a blank tab

Also, the onSecurityChange stuff was broken out into a separate
function, as was the status update stuff.  This will allow them to
easily be called from separate listeners if need be.