add new BROKEN state
This commit fixes issue #2217, where the ext would get caught in a
loop if an https site was sourcing non-https content.  This is now
hanled by new BROKEN security state handling, which is similar to
VALID, but provides a broken monkey icon to indicate, with updated
message to indicate state.